+++ react/FilterNavigator.js
@@ -90,6 +90,7 @@ const FilterNavigator = ({
   updateOnFilterSelectionOnMobile = false,
   showClearByFilter = false,
   showClearAllFilters = false,
+  showSelectedFiltersInTitle = false,
   priceRangeLayout = 'slider',
 }) => {
   const { isMobile } = useDevice()
@@ -275,7 +276,7 @@ const FilterNavigator = ({
               appliedFiltersOverview={appliedFiltersOverview}
               showClearByFilter={showClearByFilter}
               priceRangeLayout={priceRangeLayout}
-              scrollToTop={scrollToTop}
+              showSelectedFiltersInTitle={showSelectedFiltersInTitle}
             />
             {showClearAllFilters && hasFiltersApplied && (
               <div