+++ react/FilterNavigatorFlexible.js
@@ -25,6 +25,7 @@ const withSearchPageContextProps = (Component) => ({
   updateOnFilterSelectionOnMobile,
   showClearByFilter,
   showClearAllFilters,
+  showSelectedFiltersInTitle,
   priceRangeLayout,
   facetOrdering = [],
 }) => {
@@ -105,6 +106,7 @@ const withSearchPageContextProps = (Component) => ({
           updateOnFilterSelectionOnMobile={updateOnFilterSelectionOnMobile}
           showClearByFilter={showClearByFilter}
           showClearAllFilters={showClearAllFilters}
+          showSelectedFiltersInTitle={showSelectedFiltersInTitle}
           priceRangeLayout={priceRangeLayout}
         />
       </FilterNavigatorContext.Provider>